@@ -1,7 +1,7 @@
-name: hop_maldives
+name: flutteram
 
 environment:
-  sdk: '>=3.3.2 <4.0.0'
+  sdk: 3.7.2
 
 dev_dependencies:
   melos: ^3.4.0
