@@ -4,12 +4,12 @@ version: 1.0.0
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=3.3.2 <4.0.0'
+  sdk: 3.7.2
 
 dependencies:
-  stormberry: ^0.14.0
+  stormberry: 0.15.0
 
 dev_dependencies:
-  build_runner: ^2.3.3
-  lints: ^3.0.0
-  test: ^1.16.0
+  build_runner: 2.4.15
+  lints: 5.1.1
+  test: 1.25.15
