@@ -4,22 +4,21 @@ version: 1.0.0
 publish_to: none
 
 environment:
-  sdk: '>=3.3.2 <4.0.0'
+  sdk: 3.7.2
 
 dependency_overrides:
-  analyzer: ^5.3.1
+  analyzer: 5.3.1
 
 dependencies:
   db:
     path: ../db
 
-  freezed_annotation: ^2.4.1
-  json_annotation: ^4.8.0
+  freezed_annotation: 2.4.4
+  json_annotation: 4.9.0
 
 dev_dependencies:
-  build_runner: ^2.3.3
-  freezed: ^2.4.7
-  json_serializable: ^6.7.1
-  lints: ^3.0.0
-  test: ^1.16.0
+  build_runner: 2.4.15
+  freezed: 2.5.7
+  json_serializable: 6.9.0
+  lints: 5.1.1
 
