@@ -10,75 +10,83 @@ class GoogleIconPainter extends CustomPainter {
     path_0.lineTo(size.width * 0.2148846, size.height * 0.7383478);
     path_0.lineTo(size.width * 0.09628083, size.height * 0.7409087);
     path_0.cubicTo(
-        size.width * 0.06083583,
-        size.height * 0.6737783,
-        size.width * 0.04073079,
-        size.height * 0.5969696,
-        size.width * 0.04073079,
-        size.height * 0.5153478);
+      size.width * 0.06083583,
+      size.height * 0.6737783,
+      size.width * 0.04073079,
+      size.height * 0.5969696,
+      size.width * 0.04073079,
+      size.height * 0.5153478,
+    );
     path_0.cubicTo(
-        size.width * 0.04073079,
-        size.height * 0.4364174,
-        size.width * 0.05952833,
-        size.height * 0.3619878,
-        size.width * 0.09284875,
-        size.height * 0.2964504);
+      size.width * 0.04073079,
+      size.height * 0.4364174,
+      size.width * 0.05952833,
+      size.height * 0.3619878,
+      size.width * 0.09284875,
+      size.height * 0.2964504,
+    );
     path_0.lineTo(size.width * 0.09287417, size.height * 0.2964504);
     path_0.lineTo(size.width * 0.1984650, size.height * 0.3162183);
     path_0.lineTo(size.width * 0.2447200, size.height * 0.4233939);
     path_0.cubicTo(
-        size.width * 0.2350387,
-        size.height * 0.4522130,
-        size.width * 0.2297621,
-        size.height * 0.4831522,
-        size.width * 0.2297621,
-        size.height * 0.5153478);
+      size.width * 0.2350387,
+      size.height * 0.4522130,
+      size.width * 0.2297621,
+      size.height * 0.4831522,
+      size.width * 0.2297621,
+      size.height * 0.5153478,
+    );
     path_0.cubicTo(
-        size.width * 0.2297658,
-        size.height * 0.5502870,
-        size.width * 0.2359637,
-        size.height * 0.5837609,
-        size.width * 0.2473346,
-        size.height * 0.6146478);
+      size.width * 0.2297658,
+      size.height * 0.5502870,
+      size.width * 0.2359637,
+      size.height * 0.5837609,
+      size.width * 0.2473346,
+      size.height * 0.6146478,
+    );
     path_0.close();
 
     Paint paint0Fill = Paint()..style = PaintingStyle.fill;
-    paint0Fill.color = const Color(0xffFBBB00).withOpacity(1.0);
+    paint0Fill.color = const Color(0xffFBBB00).withValues(alpha: 1.0);
     canvas.drawPath(path_0, paint0Fill);
 
     Path path_1 = Path();
     path_1.moveTo(size.width * 0.9648167, size.height * 0.4264287);
     path_1.cubicTo(
-        size.width * 0.9701708,
-        size.height * 0.4552217,
-        size.width * 0.9729625,
-        size.height * 0.4849565,
-        size.width * 0.9729625,
-        size.height * 0.5153478);
+      size.width * 0.9701708,
+      size.height * 0.4552217,
+      size.width * 0.9729625,
+      size.height * 0.4849565,
+      size.width * 0.9729625,
+      size.height * 0.5153478,
+    );
     path_1.cubicTo(
-        size.width * 0.9729625,
-        size.height * 0.5494217,
-        size.width * 0.9694542,
-        size.height * 0.5826652,
-        size.width * 0.9627708,
-        size.height * 0.6147261);
+      size.width * 0.9729625,
+      size.height * 0.5494217,
+      size.width * 0.9694542,
+      size.height * 0.5826652,
+      size.width * 0.9627708,
+      size.height * 0.6147261,
+    );
     path_1.cubicTo(
-        size.width * 0.9400792,
-        size.height * 0.7238348,
-        size.width * 0.8807875,
-        size.height * 0.8191087,
-        size.width * 0.7986542,
-        size.height * 0.8865261);
+      size.width * 0.9400792,
+      size.height * 0.7238348,
+      size.width * 0.8807875,
+      size.height * 0.8191087,
+      size.width * 0.7986542,
+      size.height * 0.8865261,
+    );
     path_1.lineTo(size.width * 0.7986292, size.height * 0.8865000);
     path_1.lineTo(size.width * 0.6656333, size.height * 0.8795739);
     path_1.lineTo(size.width * 0.6468083, size.height * 0.7595826);
     path_1.cubicTo(
-        size.width * 0.7013083,
-        size.height * 0.7269478,
-        size.width * 0.7439000,
-        size.height * 0.6758696,
-        size.width * 0.7663375,
-        size.height * 0.6147261);
+      size.width * 0.7013083,
+      size.height * 0.7269478,
+      size.width * 0.7439000,
+      size.height * 0.6758696,
+      size.width * 0.7663375,
+      size.height * 0.6147261,
+    );
     path_1.lineTo(size.width * 0.5170917, size.height * 0.6147261);
     path_1.lineTo(size.width * 0.5170917, size.height * 0.4264287);
     path_1.lineTo(size.width * 0.7699708, size.height * 0.4264287);
@@ -86,85 +94,93 @@ class GoogleIconPainter extends CustomPainter {
     path_1.close();
 
     Paint paint1Fill = Paint()..style = PaintingStyle.fill;
-    paint1Fill.color = const Color(0xff518EF8).withOpacity(1.0);
+    paint1Fill.color = const Color(0xff518EF8).withValues(alpha: 1.0);
     canvas.drawPath(path_1, paint1Fill);
 
     Path path_2 = Path();
     path_2.moveTo(size.width * 0.7986292, size.height * 0.8865043);
     path_2.lineTo(size.width * 0.7986542, size.height * 0.8865304);
     path_2.cubicTo(
-        size.width * 0.7187750,
-        size.height * 0.9520957,
-        size.width * 0.6173000,
-        size.height * 0.9913217,
-        size.width * 0.5068417,
-        size.height * 0.9913217);
+      size.width * 0.7187750,
+      size.height * 0.9520957,
+      size.width * 0.6173000,
+      size.height * 0.9913217,
+      size.width * 0.5068417,
+      size.height * 0.9913217,
+    );
     path_2.cubicTo(
-        size.width * 0.3293325,
-        size.height * 0.9913217,
-        size.width * 0.1750012,
-        size.height * 0.8900087,
-        size.width * 0.09627292,
-        size.height * 0.7409130);
+      size.width * 0.3293325,
+      size.height * 0.9913217,
+      size.width * 0.1750012,
+      size.height * 0.8900087,
+      size.width * 0.09627292,
+      size.height * 0.7409130,
+    );
     path_2.lineTo(size.width * 0.2473262, size.height * 0.6146522);
     path_2.cubicTo(
-        size.width * 0.2866900,
-        size.height * 0.7219261,
-        size.width * 0.3880325,
-        size.height * 0.7982913,
-        size.width * 0.5068417,
-        size.height * 0.7982913);
+      size.width * 0.2866900,
+      size.height * 0.7219261,
+      size.width * 0.3880325,
+      size.height * 0.7982913,
+      size.width * 0.5068417,
+      size.height * 0.7982913,
+    );
     path_2.cubicTo(
-        size.width * 0.5579083,
-        size.height * 0.7982913,
-        size.width * 0.6057542,
-        size.height * 0.7841957,
-        size.width * 0.6468042,
-        size.height * 0.7595870);
+      size.width * 0.5579083,
+      size.height * 0.7982913,
+      size.width * 0.6057542,
+      size.height * 0.7841957,
+      size.width * 0.6468042,
+      size.height * 0.7595870,
+    );
     path_2.lineTo(size.width * 0.7986292, size.height * 0.8865043);
     path_2.close();
 
     Paint paint2Fill = Paint()..style = PaintingStyle.fill;
-    paint2Fill.color = const Color(0xff28B446).withOpacity(1.0);
+    paint2Fill.color = const Color(0xff28B446).withValues(alpha: 1.0);
     canvas.drawPath(path_2, paint2Fill);
 
     Path path_3 = Path();
     path_3.moveTo(size.width * 0.8043375, size.height * 0.1489483);
     path_3.lineTo(size.width * 0.6533375, size.height * 0.2751857);
     path_3.cubicTo(
-        size.width * 0.6108458,
-        size.height * 0.2480661,
-        size.width * 0.5606250,
-        size.height * 0.2324000,
-        size.width * 0.5068167,
-        size.height * 0.2324000);
+      size.width * 0.6108458,
+      size.height * 0.2480661,
+      size.width * 0.5606250,
+      size.height * 0.2324000,
+      size.width * 0.5068167,
+      size.height * 0.2324000,
+    );
     path_3.cubicTo(
-        size.width * 0.3853167,
-        size.height * 0.2324000,
-        size.width * 0.2820787,
-        size.height * 0.3122687,
-        size.width * 0.2446871,
-        size.height * 0.4233917);
+      size.width * 0.3853167,
+      size.height * 0.2324000,
+      size.width * 0.2820787,
+      size.height * 0.3122687,
+      size.width * 0.2446871,
+      size.height * 0.4233917,
+    );
     path_3.lineTo(size.width * 0.09283958, size.height * 0.2964483);
     path_3.lineTo(size.width * 0.09281417, size.height * 0.2964483);
     path_3.cubicTo(
-        size.width * 0.1703904,
-        size.height * 0.1437183,
-        size.width * 0.3266675,
-        size.height * 0.03937035,
-        size.width * 0.5068167,
-        size.height * 0.03937035);
+      size.width * 0.1703904,
+      size.height * 0.1437183,
+      size.width * 0.3266675,
+      size.height * 0.03937035,
+      size.width * 0.5068167,
+      size.height * 0.03937035,
+    );
     path_3.cubicTo(
-        size.width * 0.6199125,
-        size.height * 0.03937035,
-        size.width * 0.7236125,
-        size.height * 0.08050870,
-        size.width * 0.8043375,
-        size.height * 0.1489483);
+      size.width * 0.6199125,
+      size.height * 0.03937035,
+      size.width * 0.7236125,
+      size.height * 0.08050870,
+      size.width * 0.8043375,
+      size.height * 0.1489483,
+    );
     path_3.close();
 
     Paint paint3Fill = Paint()..style = PaintingStyle.fill;
-    paint3Fill.color = const Color(0xffF14336).withOpacity(1.0);
+    paint3Fill.color = const Color(0xffF14336).withValues(alpha: 1.0);
     canvas.drawPath(path_3, paint3Fill);
   }
 
