@@ -11,260 +11,296 @@ class PhoneIconPainter extends CustomPainter {
     Path path_0 = Path();
     path_0.moveTo(size.width * 0.8988000, size.height * 0.7637500);
     path_0.cubicTo(
-        size.width * 0.8988000,
-        size.height * 0.7787500,
-        size.width * 0.8956000,
-        size.height * 0.7941667,
-        size.width * 0.8888000,
-        size.height * 0.8091667);
+      size.width * 0.8988000,
+      size.height * 0.7787500,
+      size.width * 0.8956000,
+      size.height * 0.7941667,
+      size.width * 0.8888000,
+      size.height * 0.8091667,
+    );
     path_0.cubicTo(
-        size.width * 0.8820000,
-        size.height * 0.8241667,
-        size.width * 0.8732000,
-        size.height * 0.8383333,
-        size.width * 0.8616000,
-        size.height * 0.8516667);
+      size.width * 0.8820000,
+      size.height * 0.8241667,
+      size.width * 0.8732000,
+      size.height * 0.8383333,
+      size.width * 0.8616000,
+      size.height * 0.8516667,
+    );
     path_0.cubicTo(
-        size.width * 0.8420000,
-        size.height * 0.8741667,
-        size.width * 0.8204000,
-        size.height * 0.8904167,
-        size.width * 0.7960000,
-        size.height * 0.9008333);
+      size.width * 0.8420000,
+      size.height * 0.8741667,
+      size.width * 0.8204000,
+      size.height * 0.8904167,
+      size.width * 0.7960000,
+      size.height * 0.9008333,
+    );
     path_0.cubicTo(
-        size.width * 0.7720000,
-        size.height * 0.9112500,
-        size.width * 0.7460000,
-        size.height * 0.9166667,
-        size.width * 0.7180000,
-        size.height * 0.9166667);
+      size.width * 0.7720000,
+      size.height * 0.9112500,
+      size.width * 0.7460000,
+      size.height * 0.9166667,
+      size.width * 0.7180000,
+      size.height * 0.9166667,
+    );
     path_0.cubicTo(
-        size.width * 0.6772000,
-        size.height * 0.9166667,
-        size.width * 0.6336000,
-        size.height * 0.9066667,
-        size.width * 0.5876000,
-        size.height * 0.8862500);
+      size.width * 0.6772000,
+      size.height * 0.9166667,
+      size.width * 0.6336000,
+      size.height * 0.9066667,
+      size.width * 0.5876000,
+      size.height * 0.8862500,
+    );
     path_0.cubicTo(
-        size.width * 0.5416000,
-        size.height * 0.8658333,
-        size.width * 0.4956000,
-        size.height * 0.8383333,
-        size.width * 0.4500000,
-        size.height * 0.8037500);
+      size.width * 0.5416000,
+      size.height * 0.8658333,
+      size.width * 0.4956000,
+      size.height * 0.8383333,
+      size.width * 0.4500000,
+      size.height * 0.8037500,
+    );
     path_0.cubicTo(
-        size.width * 0.4040000,
-        size.height * 0.7687500,
-        size.width * 0.3604000,
-        size.height * 0.7300000,
-        size.width * 0.3188000,
-        size.height * 0.6870833);
+      size.width * 0.4040000,
+      size.height * 0.7687500,
+      size.width * 0.3604000,
+      size.height * 0.7300000,
+      size.width * 0.3188000,
+      size.height * 0.6870833,
+    );
     path_0.cubicTo(
-        size.width * 0.2776000,
-        size.height * 0.6437500,
-        size.width * 0.2404000,
-        size.height * 0.5983333,
-        size.width * 0.2072000,
-        size.height * 0.5508333);
+      size.width * 0.2776000,
+      size.height * 0.6437500,
+      size.width * 0.2404000,
+      size.height * 0.5983333,
+      size.width * 0.2072000,
+      size.height * 0.5508333,
+    );
     path_0.cubicTo(
-        size.width * 0.1744000,
-        size.height * 0.5033333,
-        size.width * 0.1480000,
-        size.height * 0.4558333,
-        size.width * 0.1288000,
-        size.height * 0.4087500);
+      size.width * 0.1744000,
+      size.height * 0.5033333,
+      size.width * 0.1480000,
+      size.height * 0.4558333,
+      size.width * 0.1288000,
+      size.height * 0.4087500,
+    );
     path_0.cubicTo(
-        size.width * 0.1096000,
-        size.height * 0.3612500,
-        size.width * 0.1000000,
-        size.height * 0.3158333,
-        size.width * 0.1000000,
-        size.height * 0.2725000);
+      size.width * 0.1096000,
+      size.height * 0.3612500,
+      size.width * 0.1000000,
+      size.height * 0.3158333,
+      size.width * 0.1000000,
+      size.height * 0.2725000,
+    );
     path_0.cubicTo(
-        size.width * 0.1000000,
-        size.height * 0.2441667,
-        size.width * 0.1048000,
-        size.height * 0.2170833,
-        size.width * 0.1144000,
-        size.height * 0.1920833);
+      size.width * 0.1000000,
+      size.height * 0.2441667,
+      size.width * 0.1048000,
+      size.height * 0.2170833,
+      size.width * 0.1144000,
+      size.height * 0.1920833,
+    );
     path_0.cubicTo(
-        size.width * 0.1240000,
-        size.height * 0.1666667,
-        size.width * 0.1392000,
-        size.height * 0.1433333,
-        size.width * 0.1604000,
-        size.height * 0.1225000);
+      size.width * 0.1240000,
+      size.height * 0.1666667,
+      size.width * 0.1392000,
+      size.height * 0.1433333,
+      size.width * 0.1604000,
+      size.height * 0.1225000,
+    );
     path_0.cubicTo(
-        size.width * 0.1860000,
-        size.height * 0.09625000,
-        size.width * 0.2140000,
-        size.height * 0.08333333,
-        size.width * 0.2436000,
-        size.height * 0.08333333);
+      size.width * 0.1860000,
+      size.height * 0.09625000,
+      size.width * 0.2140000,
+      size.height * 0.08333333,
+      size.width * 0.2436000,
+      size.height * 0.08333333,
+    );
     path_0.cubicTo(
-        size.width * 0.2548000,
-        size.height * 0.08333333,
-        size.width * 0.2660000,
-        size.height * 0.08583333,
-        size.width * 0.2760000,
-        size.height * 0.09083333);
+      size.width * 0.2548000,
+      size.height * 0.08333333,
+      size.width * 0.2660000,
+      size.height * 0.08583333,
+      size.width * 0.2760000,
+      size.height * 0.09083333,
+    );
     path_0.cubicTo(
-        size.width * 0.2864000,
-        size.height * 0.09583333,
-        size.width * 0.2956000,
-        size.height * 0.1033333,
-        size.width * 0.3028000,
-        size.height * 0.1141667);
+      size.width * 0.2864000,
+      size.height * 0.09583333,
+      size.width * 0.2956000,
+      size.height * 0.1033333,
+      size.width * 0.3028000,
+      size.height * 0.1141667,
+    );
     path_0.lineTo(size.width * 0.3956000, size.height * 0.2504167);
     path_0.cubicTo(
-        size.width * 0.4028000,
-        size.height * 0.2608333,
-        size.width * 0.4080000,
-        size.height * 0.2704167,
-        size.width * 0.4116000,
-        size.height * 0.2795833);
+      size.width * 0.4028000,
+      size.height * 0.2608333,
+      size.width * 0.4080000,
+      size.height * 0.2704167,
+      size.width * 0.4116000,
+      size.height * 0.2795833,
+    );
     path_0.cubicTo(
-        size.width * 0.4152000,
-        size.height * 0.2883333,
-        size.width * 0.4172000,
-        size.height * 0.2970833,
-        size.width * 0.4172000,
-        size.height * 0.3050000);
+      size.width * 0.4152000,
+      size.height * 0.2883333,
+      size.width * 0.4172000,
+      size.height * 0.2970833,
+      size.width * 0.4172000,
+      size.height * 0.3050000,
+    );
     path_0.cubicTo(
-        size.width * 0.4172000,
-        size.height * 0.3150000,
-        size.width * 0.4144000,
-        size.height * 0.3250000,
-        size.width * 0.4088000,
-        size.height * 0.3345833);
+      size.width * 0.4172000,
+      size.height * 0.3150000,
+      size.width * 0.4144000,
+      size.height * 0.3250000,
+      size.width * 0.4088000,
+      size.height * 0.3345833,
+    );
     path_0.cubicTo(
-        size.width * 0.4036000,
-        size.height * 0.3441667,
-        size.width * 0.3960000,
-        size.height * 0.3541667,
-        size.width * 0.3864000,
-        size.height * 0.3641667);
+      size.width * 0.4036000,
+      size.height * 0.3441667,
+      size.width * 0.3960000,
+      size.height * 0.3541667,
+      size.width * 0.3864000,
+      size.height * 0.3641667,
+    );
     path_0.lineTo(size.width * 0.3560000, size.height * 0.3970833);
     path_0.cubicTo(
-        size.width * 0.3516000,
-        size.height * 0.4016667,
-        size.width * 0.3496000,
-        size.height * 0.4070833,
-        size.width * 0.3496000,
-        size.height * 0.4137500);
+      size.width * 0.3516000,
+      size.height * 0.4016667,
+      size.width * 0.3496000,
+      size.height * 0.4070833,
+      size.width * 0.3496000,
+      size.height * 0.4137500,
+    );
     path_0.cubicTo(
-        size.width * 0.3496000,
-        size.height * 0.4170833,
-        size.width * 0.3500000,
-        size.height * 0.4200000,
-        size.width * 0.3508000,
-        size.height * 0.4233333);
+      size.width * 0.3496000,
+      size.height * 0.4170833,
+      size.width * 0.3500000,
+      size.height * 0.4200000,
+      size.width * 0.3508000,
+      size.height * 0.4233333,
+    );
     path_0.cubicTo(
-        size.width * 0.3520000,
-        size.height * 0.4266667,
-        size.width * 0.3532000,
-        size.height * 0.4291667,
-        size.width * 0.3540000,
-        size.height * 0.4316667);
+      size.width * 0.3520000,
+      size.height * 0.4266667,
+      size.width * 0.3532000,
+      size.height * 0.4291667,
+      size.width * 0.3540000,
+      size.height * 0.4316667,
+    );
     path_0.cubicTo(
-        size.width * 0.3612000,
-        size.height * 0.4454167,
-        size.width * 0.3736000,
-        size.height * 0.4633333,
-        size.width * 0.3912000,
-        size.height * 0.4850000);
+      size.width * 0.3612000,
+      size.height * 0.4454167,
+      size.width * 0.3736000,
+      size.height * 0.4633333,
+      size.width * 0.3912000,
+      size.height * 0.4850000,
+    );
     path_0.cubicTo(
-        size.width * 0.4092000,
-        size.height * 0.5066667,
-        size.width * 0.4284000,
-        size.height * 0.5287500,
-        size.width * 0.4492000,
-        size.height * 0.5508333);
+      size.width * 0.4092000,
+      size.height * 0.5066667,
+      size.width * 0.4284000,
+      size.height * 0.5287500,
+      size.width * 0.4492000,
+      size.height * 0.5508333,
+    );
     path_0.cubicTo(
-        size.width * 0.4708000,
-        size.height * 0.5729167,
-        size.width * 0.4916000,
-        size.height * 0.5933333,
-        size.width * 0.5128000,
-        size.height * 0.6120833);
+      size.width * 0.4708000,
+      size.height * 0.5729167,
+      size.width * 0.4916000,
+      size.height * 0.5933333,
+      size.width * 0.5128000,
+      size.height * 0.6120833,
+    );
     path_0.cubicTo(
-        size.width * 0.5336000,
-        size.height * 0.6304167,
-        size.width * 0.5508000,
-        size.height * 0.6429167,
-        size.width * 0.5644000,
-        size.height * 0.6504167);
+      size.width * 0.5336000,
+      size.height * 0.6304167,
+      size.width * 0.5508000,
+      size.height * 0.6429167,
+      size.width * 0.5644000,
+      size.height * 0.6504167,
+    );
     path_0.cubicTo(
-        size.width * 0.5664000,
-        size.height * 0.6512500,
-        size.width * 0.5688000,
-        size.height * 0.6525000,
-        size.width * 0.5716000,
-        size.height * 0.6537500);
+      size.width * 0.5664000,
+      size.height * 0.6512500,
+      size.width * 0.5688000,
+      size.height * 0.6525000,
+      size.width * 0.5716000,
+      size.height * 0.6537500,
+    );
     path_0.cubicTo(
-        size.width * 0.5748000,
-        size.height * 0.6550000,
-        size.width * 0.5780000,
-        size.height * 0.6554167,
-        size.width * 0.5816000,
-        size.height * 0.6554167);
+      size.width * 0.5748000,
+      size.height * 0.6550000,
+      size.width * 0.5780000,
+      size.height * 0.6554167,
+      size.width * 0.5816000,
+      size.height * 0.6554167,
+    );
     path_0.cubicTo(
-        size.width * 0.5884000,
-        size.height * 0.6554167,
-        size.width * 0.5936000,
-        size.height * 0.6529167,
-        size.width * 0.5980000,
-        size.height * 0.6483333);
+      size.width * 0.5884000,
+      size.height * 0.6554167,
+      size.width * 0.5936000,
+      size.height * 0.6529167,
+      size.width * 0.5980000,
+      size.height * 0.6483333,
+    );
     path_0.lineTo(size.width * 0.6284000, size.height * 0.6170833);
     path_0.cubicTo(
-        size.width * 0.6384000,
-        size.height * 0.6066667,
-        size.width * 0.6480000,
-        size.height * 0.5987500,
-        size.width * 0.6572000,
-        size.height * 0.5937500);
+      size.width * 0.6384000,
+      size.height * 0.6066667,
+      size.width * 0.6480000,
+      size.height * 0.5987500,
+      size.width * 0.6572000,
+      size.height * 0.5937500,
+    );
     path_0.cubicTo(
-        size.width * 0.6664000,
-        size.height * 0.5879167,
-        size.width * 0.6756000,
-        size.height * 0.5850000,
-        size.width * 0.6856000,
-        size.height * 0.5850000);
+      size.width * 0.6664000,
+      size.height * 0.5879167,
+      size.width * 0.6756000,
+      size.height * 0.5850000,
+      size.width * 0.6856000,
+      size.height * 0.5850000,
+    );
     path_0.cubicTo(
-        size.width * 0.6932000,
-        size.height * 0.5850000,
-        size.width * 0.7012000,
-        size.height * 0.5866667,
-        size.width * 0.7100000,
-        size.height * 0.5904167);
+      size.width * 0.6932000,
+      size.height * 0.5850000,
+      size.width * 0.7012000,
+      size.height * 0.5866667,
+      size.width * 0.7100000,
+      size.height * 0.5904167,
+    );
     path_0.cubicTo(
-        size.width * 0.7188000,
-        size.height * 0.5941667,
-        size.width * 0.7280000,
-        size.height * 0.5995833,
-        size.width * 0.7380000,
-        size.height * 0.6066667);
+      size.width * 0.7188000,
+      size.height * 0.5941667,
+      size.width * 0.7280000,
+      size.height * 0.5995833,
+      size.width * 0.7380000,
+      size.height * 0.6066667,
+    );
     path_0.lineTo(size.width * 0.8704000, size.height * 0.7045833);
     path_0.cubicTo(
-        size.width * 0.8808000,
-        size.height * 0.7120833,
-        size.width * 0.8880000,
-        size.height * 0.7208333,
-        size.width * 0.8924000,
-        size.height * 0.7312500);
+      size.width * 0.8808000,
+      size.height * 0.7120833,
+      size.width * 0.8880000,
+      size.height * 0.7208333,
+      size.width * 0.8924000,
+      size.height * 0.7312500,
+    );
     path_0.cubicTo(
-        size.width * 0.8964000,
-        size.height * 0.7416667,
-        size.width * 0.8988000,
-        size.height * 0.7520833,
-        size.width * 0.8988000,
-        size.height * 0.7637500);
+      size.width * 0.8964000,
+      size.height * 0.7416667,
+      size.width * 0.8988000,
+      size.height * 0.7520833,
+      size.width * 0.8988000,
+      size.height * 0.7637500,
+    );
     path_0.close();
 
-    Paint paint0Stroke = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = size.width * 0.06000000;
-    paint0Stroke.color = Colors.white.withOpacity(1.0);
+    Paint paint0Stroke =
+        Paint()
+          ..style = PaintingStyle.stroke
+          ..strokeWidth = size.width * 0.06000000;
+    paint0Stroke.color = Colors.white.withValues(alpha: 1.0);
     canvas.drawPath(path_0, paint0Stroke);
 
     Paint paint0Fill = Paint()..style = PaintingStyle.fill;
