@@ -1,14 +1,17 @@
 part of 'notification.dart';
 
 class LocalNotification {
+  static late GlobalKey<NavigatorState> _navigator;
   static final ins = FlutterLocalNotificationsPlugin();
 
-  static void init() async {
+  static void init(GlobalKey<NavigatorState> state) async {
+    _navigator = state;
     const initializationSettings = InitializationSettings(
-      android: AndroidInitializationSettings('@drawable/ic_nofitication'),
+      android: AndroidInitializationSettings('@drawable/ic_notify'),
       iOS: DarwinInitializationSettings(
-        onDidReceiveLocalNotification:
-            LocalNotification.onDidReceiveLocalNotification,
+        requestAlertPermission: true,
+        requestBadgePermission: true,
+        requestSoundPermission: true,
       ),
     );
 
@@ -22,20 +25,42 @@ class LocalNotification {
     );
 
     if (Platform.isAndroid) {
-      final hasAndroidPermission = await LocalNotification.ins
-          .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
-          ?.requestNotificationsPermission();
-      debugPrint(
-          '--- Android 13 and above has notification permission ----- $hasAndroidPermission');
+      final hasAndroidPermission =
+          await LocalNotification.ins
+              .resolvePlatformSpecificImplementation<
+                AndroidFlutterLocalNotificationsPlugin
+              >()
+              ?.requestNotificationsPermission();
+      AppLog.log(
+        '--- Android 13 and above has notification permission ----- $hasAndroidPermission',
+      );
     }
   }
 
-  static Future handleNotificationAction(Map<String, dynamic> payload) async {
+  static Future<void> handleNotificationAction(
+    Map<String, dynamic> payload,
+  ) async {
     if (payload.isNotEmpty) {
-      // do something
+      // final String type = payload['type'] ?? '';
+      // final sessionId = payload['sessionId'] as String?;
+
+      // if (type == NotifyType.session.name) {
+      //   Navigator.pushNamedAndRemoveUntil(
+      //     _navigator.currentContext!,
+      //     AppRoutes.sessionDetails,
+      //     arguments: {'sessionId': sessionId},
+      //     (route) => false,
+      //   );
+
+      //   return;
+      // }
+      // Navigator.pushNamedAndRemoveUntil(
+      //   _navigator.currentContext!,
+      //   AppRoutes.appNotifications,
+      //   (route) => false,
+      // );
     } else {
-      return AppAlice.showInspector();
+      AppAlice.showInspector();
     }
   }
 
@@ -54,10 +79,9 @@ class LocalNotification {
           priority: Priority.high,
           importance: Importance.max,
           playSound: true,
+          styleInformation: BigTextStyleInformation(notification.body ?? ''),
         ),
-        iOS: const DarwinNotificationDetails(
-          presentSound: true,
-        ),
+        iOS: const DarwinNotificationDetails(presentSound: true),
       ),
       payload: jsonEncode(message.data),
     );
