@@ -23,7 +23,9 @@ class _CoverPhoto extends StatelessWidget {
         return GestureDetector(
           onTap: () async {
             if (state.cover is CoverUploadLoading ||
-                mediaCubit.state is UploadCoverLoading) return;
+                mediaCubit.state is UploadCoverLoading) {
+              return;
+            }
 
             final hasCover = user.coverURL.isNotEmpty;
             final value = await showModalBottomSheet(
@@ -52,80 +54,77 @@ class _CoverPhoto extends StatelessWidget {
               mediaCubit.removeMedia(user.coverURL);
             }
           },
-          child: hasCoverPhoto &&
-                  state.cover is! CoverUploadLoading &&
-                  mediaCubit.state is! UploadCoverLoading
-              ? Container(
-                  height: 80.un(),
-                  width: double.infinity,
-                  decoration: BoxDecoration(
-                    borderRadius: 12.radius(),
-                    border: Border.all(
-                      color: AppTheme.primary,
-                      width: 1.un(),
-                    ),
-                  ),
-                  child: Stack(
-                    children: [
-                      Positioned.fill(
-                        child: ClipRRect(
-                          borderRadius: 12.radius(),
-                          child: CachedNetworkImage(
-                            imageUrl: user.coverURL,
-                            fit: BoxFit.cover,
-                          ),
-                        ),
+          child:
+              hasCoverPhoto &&
+                      state.cover is! CoverUploadLoading &&
+                      mediaCubit.state is! UploadCoverLoading
+                  ? Container(
+                    height: 80.un(),
+                    width: double.infinity,
+                    decoration: BoxDecoration(
+                      borderRadius: 12.radius(),
+                      border: Border.all(
+                        color: AppTheme.primary,
+                        width: 1.un(),
                       ),
-                      Positioned(
-                        bottom: 10,
-                        right: 10,
-                        child: SizedBox(
-                          height: 20.un(),
-                          child: const AppIconButton(
-                            color: AppTheme.primary,
-                            icon: Icon(
-                              Iconsax.gallery_add,
+                    ),
+                    child: Stack(
+                      children: [
+                        Positioned.fill(
+                          child: ClipRRect(
+                            borderRadius: 12.radius(),
+                            child: CachedNetworkImage(
+                              imageUrl: user.coverURL,
+                              fit: BoxFit.cover,
                             ),
-                            onTap: null,
                           ),
                         ),
-                      )
-                    ],
-                  ),
-                )
-              : DottedBorder(
-                  borderType: BorderType.RRect,
-                  radius: const Radius.circular(12),
-                  strokeWidth: 2,
-                  dashPattern: const [10],
-                  color: AppTheme.grey,
-                  child: Center(
-                    child: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        Space.y.t100,
-                        state.cover is CoverUploadLoading ||
-                                mediaCubit.state is UploadCoverLoading
-                            ? const CircularProgressIndicator(
-                                color: Colors.white,
-                              )
-                            : CustomPaint(
-                                painter: const GalleryIconPainter(),
-                                size: GalleryIconPainter.s(
-                                  15.un(),
-                                ),
-                              ),
-                        Space.y.t20,
-                        Text(
-                          'Upload cover photo',
-                          style: AppText.b3 + AppTheme.grey,
+                        Positioned(
+                          bottom: 10,
+                          right: 10,
+                          child: SizedBox(
+                            height: 20.un(),
+                            child: const AppIconButton(
+                              color: AppTheme.primary,
+                              icon: Icon(Iconsax.gallery_add),
+                              onTap: null,
+                            ),
+                          ),
                         ),
-                        Space.y.t100,
                       ],
                     ),
+                  )
+                  : DottedBorder(
+                    borderType: BorderType.RRect,
+                    radius: const Radius.circular(12),
+                    strokeWidth: 2,
+                    dashPattern: const [10],
+                    color: AppTheme.grey,
+                    child: Center(
+                      child: Column(
+                        mainAxisSize: MainAxisSize.min,
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: [
+                          Space.y.t100,
+                          state.cover is CoverUploadLoading ||
+                                  mediaCubit.state is UploadCoverLoading
+                              ? const CircularProgressIndicator(
+                                color: Colors.white,
+                              )
+                              : CustomPaint(
+                                painter: const GalleryIconPainter(),
+                                size: GalleryIconPainter.s(15.un()),
+                              ),
+                          Space.y.t20,
+                          Text(
+                            'Upload cover photo',
+                            style: AppText.b3 + AppTheme.grey,
+                          ),
+                          Space.y.t100,
+                        ],
+                      ),
+                    ),
                   ),
-                ),
         );
       },
     );
